python-app/requirements.txt
fastapi
uvicorn
tweepy
pymongo
python-dotenv
requests
starlette
pydantic
itsdangerous
